Import and Export Runway functionality
Added an id to both the Import and Runway button in the GUI.fxml file

In the Controller.java file I created an event for both of those buttons
and added functionality to both of them. Import button imports an xml
file to the project and the export one exports it to an xml file.